The category component created
