+++ react-webshop/src/App.jsx
@@ -1,0 +1,43 @@
+import React from 'react';
+import './categories.styles.scss';
+import './components/category-item/Category-item.component.jsx'
+import CategoryItem from './components/category-item/Category-item.component.jsx';
+
+const App = () => {
+  const categories = [
+    {
+      "id": 1,
+      "title": "hats",
+      "imageUrl": "https://i.ibb.co/cvpntL1/hats.png"
+    },
+    {
+      "id": 2,
+      "title": "jackets",
+      "imageUrl": "https://i.ibb.co/px2tCc3/jackets.png"
+    },
+    {
+      "id": 3,
+      "title": "sneakers",
+      "imageUrl": "https://i.ibb.co/0jqHpnp/sneakers.png"
+    },
+    {
+      "id": 4,
+      "title": "womens",
+      "imageUrl": "https://i.ibb.co/GCCdy8t/womens.png"
+    },
+    {
+      "id": 5,
+      "title": "mens",
+      "imageUrl": "https://i.ibb.co/R70vBrQ/men.png"
+    }
+  ];
+  return (
+    <div className='categories-container'>
+      {categories.map((category) => (
+        <CategoryItem category ={category} key={category.id} />
+      ))}
+    </div>
+  );
+};
+
+export default App;
