+++ react-webshop/src/components/category-item/Category-item.component.jsx
@@ -1,0 +1,20 @@
+// import React from 'react';
+import './category-item.styles.scss';
+
+export default function CategoryItem ({category}) {
+  const { imageUrl, title } = category;
+  return (
+    <div className='category-container'>
+      <div
+        className='background-image'
+        style={{
+          backgroundImage: `url(${imageUrl})`,
+        }}
+      />
+      <div className='category-body-container'>
+        <h2>{title}</h2>
+        <p>Shop Now</p>
+      </div>
+    </div>
+  );
+}
